Fix button shadows
Buttons should not have any static text shadows by default,
because it does not work for dark Gtk themes.
This patch disables the white (hardcoded) text shadow
and adds textstyle support for the prelight state on buttons.
To configure text shadows the textstyle and text_shade
settings should be used.