Add rudimentary draw-border style property support
A border is rendered only if at least one side of the border
(specified by the draw-border style property) is >0.

Limitation: murrine draws always a 1px border on all sides.
This patch allows it to enable/disable borders, but does not
add support for borders with different sizes.

Implemented widgets: GtkButton, GtkScrollbar, GtkProgressBar