Fix ComboBox arrowtype 0 alignemnt and sizing
murrine_draw_combo_arrow assumes that the size
and coordinates are relative to the widgets metrics
which is not the case. Using relative size calculations
makes only sense if the metrics passed in are
calculated dynamically too. It would be a great idea
to rewrite the arrow alignment calculation from
scratch, but removing the calculations and making
the size static (as everywhere else) is good
enough for now to make it work as intended.